pyproject.toml
[project]
name = "task_management"
version = "0.1.0"
description = ""
authors = [
    {name = "Rishav",email = "rishishah907@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "openai (==0.28)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
